migrate test_backpack_configs to new regime
